move git utils stuff to git command
